+++ system/src/Grav/Common/Flex/FlexObject.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * @package    Grav\Common\Flex
+ *
+ * @copyright  Copyright (C) 2015 - 2020 Trilby Media, LLC. All rights reserved.
+ * @license    MIT License; see LICENSE file for details.
+ */
+
+namespace Grav\Common\Flex;
+
+use Grav\Common\Flex\Traits\FlexGravTrait;
+use Grav\Common\Flex\Traits\FlexObjectTrait;
+use Grav\Framework\Flex\Traits\FlexMediaTrait;
+
+/**
+ * Class FlexObject
+ *
+ * @package Grav\Common\Flex
+ */
+abstract class FlexObject extends \Grav\Framework\Flex\FlexObject
+{
+    use FlexGravTrait;
+    use FlexObjectTrait;
+    use FlexMediaTrait;
+}